Adds Chrome extension to WebDriver and verifies installation.
This adds the Chrome extension using the `--load-extension` argument, which makes Chrome load an "unpacked" extension. This saves us the effort of generating a CRX file. I also had to remove the headless option, as it seems headless does not support Chrome extensions like this.

I identified the extension ID by simply installing it locally and then copy-pasting the value from `chrome://extensions`. I'm not totally clear under what circumstances it could change for a dev build, but this will hopefully be good enough for a proof-of-concept.