Attempts (and fails) to access Chrome DevTools via WebDriver.
Using `--auto-open-devtools-for-tabs` makes Chrome automatically open DevTools docked in the tab. Unfortunately, that's about as far as I could get. The `getDevTools()` method doesn't actually seem to return anything useful. I believe this is actually referring to the Chrome DevTools Protocol, rather than the DevTools UI.

I attempted a few hacks that didn't go anywhere.
* I tried using `Ctrl+[` to move to the previous panel in DevTools (which would wrap around to the final panel, the extension's). However it seems that DevTools isn't focused and does not receive or react to this shortcut.
* I tried listing all windows and tabs, but they do not appear to contain the DevTools windows.
* I tried using `Ctrl+Shift+D` to toggle the DevTools window into an undocked state, where it might exist as another window. However this only toggles with the last state, so it ends up just toggling between right aligned and bottom aligned, but always docked in the same window.

I don't know enough about WebDriver to hack further on this, but AFAICT, this is currently not supported and may not be possible at all.